requirements.txt
streamlit
PyMuPDF
chromadb
google-generativeai
sentence-transformers
pandas
numpy
typing-extensions
pydantic